Organize Transfers store actions
