+++ ppr-ui/src/store/actions/mhr-transfer-actions-model.ts
@@ -1,4 +1,24 @@
-import { ActionIF, SubmittingPartyIF } from '@/interfaces'
+import { ActionIF, MhrRegistrationHomeOwnerGroupIF, MhrTransferIF, SubmittingPartyIF } from '@/interfaces'
+
+export const setEmptyMhrTransfer: ActionIF = ({ commit }, emptyMhrTransfer: MhrTransferIF): void => {
+  commit('mutateEmptyMhrTransfer', emptyMhrTransfer)
+}
+
+export const setMhrTransferHomeOwnerGroups: ActionIF = (
+  { commit },
+  groups: MhrRegistrationHomeOwnerGroupIF[]
+): void => {
+  commit('mutateMhrTransferHomeOwnerGroups', groups)
+  commit('mutateUnsavedChanges', true)
+}
+
+// Set a snapshot of the MH Registration home owner groups
+export const setMhrTransferCurrentHomeOwnerGroups: ActionIF = (
+  { commit },
+  groups: MhrRegistrationHomeOwnerGroupIF[]
+): void => {
+  commit('mutateMhrTransferCurrentHomeOwnerGroups', groups)
+}
 
 export const setMhrTransferDeclaredValue: ActionIF = ({ commit }, declaredValue: number): void => {
   commit('mutateMhrTransferDeclaredValue', declaredValue)
