+++ ppr-ui/src/store/getters/mhr-transfer-state-getters.ts
@@ -1,5 +1,34 @@
-import { StateIF, SubmittingPartyIF } from '@/interfaces'
+import { MhrRegistrationHomeOwnerGroupIF, MhrRegistrationHomeOwnerIF, StateIF, SubmittingPartyIF } from '@/interfaces'
 
+// Home Owners
+export const getMhrTransferHomeOwners = (state: StateIF): MhrRegistrationHomeOwnerIF[] => {
+  const owners = []
+  state.stateModel.mhrTransfer.ownerGroups.forEach(group => {
+    if (group.owners.length === 0) {
+      // Groups with no owners should have at least one 'placeholder' owner
+      // to be properly displayed in Group Table
+      owners.push({ groupId: group.groupId })
+    } else {
+      group.owners.forEach(owner => owners.push({ ...owner, groupId: group.groupId }))
+    }
+  })
+  return owners
+}
+
+export const getMhrTransferCurrentHomeOwners = (state: StateIF): MhrRegistrationHomeOwnerGroupIF[] => {
+  const ownerGroups = []
+
+  state.stateModel.mhrTransfer.currentOwnerGroups.forEach(group => {
+    ownerGroups.push(group)
+  })
+  return ownerGroups
+}
+
+export const getMhrTransferHomeOwnerGroups = (state: StateIF): MhrRegistrationHomeOwnerGroupIF[] => {
+  return state.stateModel.mhrTransfer.ownerGroups
+}
+
+// Ownership Transfers
 export const getMhrTransferDeclaredValue = (state: StateIF): number => {
   return state.stateModel.mhrTransfer.declaredValue
 }
