+++ ppr-ui/src/store/getters/state-getters.ts
@@ -731,36 +731,8 @@ export const getMhrRegistrationValidationModel = (state: StateIF): MhrValidation
   return state.stateModel.mhrValidationState
 }
 
-// Mhr Transfers
 export const getMhrInformation = (state: StateIF): MhRegistrationSummaryIF => {
   return state.stateModel.mhrInformation
-}
-
-export const getMhrTransferHomeOwners = (state: StateIF): MhrRegistrationHomeOwnerIF[] => {
-  const owners = []
-  state.stateModel.mhrTransfer.ownerGroups.forEach(group => {
-    if (group.owners.length === 0) {
-      // Groups with no owners should have at least one 'placeholder' owner
-      // to be properly displayed in Group Table
-      owners.push({ groupId: group.groupId })
-    } else {
-      group.owners.forEach(owner => owners.push({ ...owner, groupId: group.groupId }))
-    }
-  })
-  return owners
-}
-
-export const getMhrTransferCurrentHomeOwners = (state: StateIF): MhrRegistrationHomeOwnerGroupIF[] => {
-  const ownerGroups = []
-
-  state.stateModel.mhrTransfer.currentOwnerGroups.forEach(group => {
-    ownerGroups.push(group)
-  })
-  return ownerGroups
-}
-
-export const getMhrTransferHomeOwnerGroups = (state: StateIF): MhrRegistrationHomeOwnerGroupIF[] => {
-  return state.stateModel.mhrTransfer.ownerGroups
 }
 
 export const hasLien = (state: StateIF): boolean => {
