+++ ppr-ui/src/store/mutations/mutations-model.ts
@@ -27,7 +27,6 @@ import {
   ManufacturedHomeSearchResultIF,
   UserProductSubscriptionIF,
   MhrRegistrationHomeOwnerGroupIF,
-  MhrTransferIF,
   MhrRegistrationIF,
   MhRegistrationSummaryIF,
   SubmittingPartyIF
@@ -420,23 +419,4 @@ export const mutateMhrInformation = (state: StateIF, mhrInfo: MhRegistrationSumm
 
 export const mutateLienType = (state: StateIF, lienType: string) => {
   state.stateModel.mhrInformation.lienRegistrationType = lienType
-}
-
-// MHR Transfer
-export const mutateEmptyMhrTransfer = (state: StateIF, emptyMhrTransfer: MhrTransferIF) => {
-  state.stateModel.mhrTransfer = emptyMhrTransfer
-}
-
-export const mutateMhrTransferHomeOwnerGroups = (
-  state: StateIF,
-  groups: Array<MhrRegistrationHomeOwnerGroupIF>
-) => {
-  state.stateModel.mhrTransfer.ownerGroups = groups
-}
-
-export const mutateMhrTransferCurrentHomeOwnerGroups = (
-  state: StateIF,
-  groups: Array<MhrRegistrationHomeOwnerGroupIF>
-) => {
-  state.stateModel.mhrTransfer.currentOwnerGroups = groups
 }
